+++ integration_test/cases/browser/screenshot_test.exs
@@ -1,5 +1,6 @@
 defmodule Wallaby.Integration.Browser.ScreenshotTest do
   use Wallaby.Integration.SessionCase, async: false
+  import ExUnit.CaptureIO
 
   import Wallaby.Query, only: [css: 1]
 
@@ -78,9 +79,13 @@ defmodule Wallaby.Integration.Browser.ScreenshotTest do
 
     Application.put_env(:wallaby, :screenshot_on_failure, true)
 
-    assert_raise Wallaby.QueryError, fn ->
-      find(page, css(".some-selector"))
-    end
+    output = capture_io( fn ->
+      assert_raise Wallaby.QueryError, fn ->
+        find(page, css(".some-selector"))
+      end
+    end)
+    assert Regex.match?(~r/Screenshot taken/, output)
+
     assert File.exists?("#{File.cwd!}/screenshots")
     assert File.ls!("#{File.cwd!}/screenshots") |> Enum.count == 1