+++ integration_test/tests.exs
@@ -10,6 +10,7 @@ Code.require_file "cases/browser/click_button_test.exs", __DIR__
 Code.require_file "cases/browser/click_test.exs", __DIR__
 Code.require_file "cases/browser/cookies_test.exs", __DIR__
 Code.require_file "cases/browser/current_path_test.exs", __DIR__
+Code.require_file "cases/browser/dialog_test.exs", __DIR__
 Code.require_file "cases/browser/execute_script_test.exs", __DIR__
 Code.require_file "cases/browser/fill_in_test.exs", __DIR__
 Code.require_file "cases/browser/find_test.exs", __DIR__
@@ -19,6 +20,7 @@ Code.require_file "cases/browser/has_value_test.exs", __DIR__
 Code.require_file "cases/browser/local_storage_test.exs", __DIR__
 Code.require_file "cases/browser/navigation_test.exs", __DIR__
 Code.require_file "cases/browser/page_source_test.exs", __DIR__
+Code.require_file "cases/browser/screenshot_test.exs", __DIR__
 Code.require_file "cases/browser/select_test.exs", __DIR__
 Code.require_file "cases/browser/set_value_test.exs", __DIR__
 Code.require_file "cases/browser/send_text_test.exs", __DIR__