+++ lib/wallaby/element.ex
@@ -97,6 +97,18 @@ defmodule Wallaby.Element do
   end
 
   @doc """
+  Hovers on the element.
+  """
+  @spec hover(t) :: t
+
+  def hover(%__MODULE__{driver: driver} = element) do
+    case driver.hover(element) do
+      {:ok, _} ->
+        element
+    end
+  end
+
+  @doc """
   Returns the text from the element.
   """
   @spec text(t) :: String.t
@@ -197,5 +209,20 @@ defmodule Wallaby.Element do
 
   def value(element) do
     attr(element, "value")
+  end
+end
+
+defimpl Inspect, for: Wallaby.Element  do
+  import Inspect.Algebra
+
+  def inspect(element, opts) do
+    outer_html = Wallaby.Element.attr(element, "outerHTML")
+
+    concat([
+      Inspect.Any.inspect(element, opts),
+      "\n\n",
+      IO.ANSI.cyan <> "outerHTML:\n\n" <> IO.ANSI.reset,
+      IO.ANSI.yellow <> outer_html <> IO.ANSI.reset
+    ])
   end
 end