+++ lib/wallaby/httpclient.ex
@@ -7,6 +7,7 @@ defmodule Wallaby.HTTPClient do
   @type request_opts :: {:encode_json, boolean}
 
   @status_obscured 13
+  @max_jitter 50 # The maximum time we'll sleep is for 50ms
 
   @doc """
   Sends a request to the webdriver API and parses the
@@ -41,6 +42,7 @@ defmodule Wallaby.HTTPClient do
     |> handle_response
     |> case do
       {:error, :httpoison, error} ->
+        :timer.sleep(jitter())
         make_request(method, url, body, retry_count + 1, [inspect(error) | retry_reasons])
 
       result ->
@@ -67,17 +69,24 @@ defmodule Wallaby.HTTPClient do
   defp check_status(response) do
     case Map.get(response, "status") do
       @status_obscured ->
-        {:error, :obscured}
+        message = get_in(response, ["value", "message"])
+
+        {:error, message}
       _  ->
         {:ok, response}
     end
   end
 
+  # credo:disable-for-next-line Credo.Check.Refactor.CyclomaticComplexity
   defp check_for_response_errors(response) do
     case Map.get(response, "value") do
       %{"class" => "org.openqa.selenium.StaleElementReferenceException"} ->
         {:error, :stale_reference}
+      %{"message" => "Stale element reference" <> _} ->
+        {:error, :stale_reference}
       %{"message" => "stale element reference" <> _} ->
+        {:error, :stale_reference}
+      %{"message" => "An element command failed because the referenced element is no longer available" <> _} ->
         {:error, :stale_reference}
       %{"message" => "invalid selector" <> _} ->
         {:error, :invalid_selector}
@@ -107,4 +116,6 @@ defmodule Wallaby.HTTPClient do
   def to_params({:css, css}) do
     %{using: "css selector", value: css}
   end
+
+  defp jitter, do: :rand.uniform(@max_jitter)
 end