+++ lib/wallaby/query/error_message.ex
@@ -78,8 +78,20 @@ defmodule Wallaby.Query.ErrorMessage do
 
   defp found_error_message(query) do
     """
-    #{expected_count(query)}, #{visibility(query)} #{method(query)} '#{query.selector}' but #{result_count(query.result)}, #{visibility(query)} #{short_method(query.method, Enum.count(query.result))} #{result_expectation(query.result)}.
+    #{expected_count(query)}, #{visibility_and_selection(query)} #{method(query)} #{selector(query)} but #{result_count(query.result)}, #{visibility_and_selection(query)} #{short_method(query.method, Enum.count(query.result))} #{result_expectation(query.result)}.
     """
+  end
+
+  @doc """
+  Extracts the selector from the query
+  """
+  @spec selector(Query.t) :: String.t
+
+  def selector(%Query{selector: {name, value}}) do
+    "'#{name}' with value '#{value}'"
+  end
+  def selector(%Query{selector: selector}) do
+    "'#{selector}'"
   end
 
   @doc """
@@ -127,6 +139,9 @@ defmodule Wallaby.Query.ErrorMessage do
   def method(:text, true), do: "elements with the text"
   def method(:text, false), do: "element with the text"
 
+  def method(:attribute, true), do: "elements with the attribute"
+  def method(:attribute, false), do: "element with the attribute"
+
   def short_method(:css, count) when count > 1,  do: "elements"
   def short_method(:css, count) when count == 0, do: "elements"
   def short_method(:css, _),                 do: "element"
@@ -159,16 +174,25 @@ defmodule Wallaby.Query.ErrorMessage do
   end
   def condition(_), do: nil
 
+  @spec visibility_and_selection(Query.t) :: String.t
+  defp visibility_and_selection(query) do
+    case Query.selected?(query) do
+      true -> "#{visibility(query)}, selected"
+      false -> "#{visibility(query)}, unselected"
+      :any -> visibility(query)
+    end
+  end
+
   @doc """
   Converts the visibilty attribute into a human readable form.
   """
   @spec visibility(Query.t) :: String.t
 
   def visibility(query) do
-    if Query.visible?(query) do
-      "visible"
-    else
-      "invisible"
+    case Query.visible?(query) do
+      true -> "visible"
+      false -> "invisible"
+      :any -> "visible or invisible"
     end
   end